+++ CHANGELOG.md
@@ -12,7 +12,8 @@ All notable changes will be documented in this file
 * [PR-66](https://github.com/pablobascunana/youml-api/pull/66) Training model
 
 ### Changed
-* [PR-84](https://github.com/pablobascunana/youml-api/pull/72) Rename some files references from youml-manager to youml-api
+* [PR-85](https://github.com/pablobascunana/youml-api/pull/85) Bump tox version from 4.0.11 to 4.0.18
+* [PR-84](https://github.com/pablobascunana/youml-api/pull/84) Rename some files references from youml-manager to youml-api
 * [PR-69](https://github.com/pablobascunana/youml-api/pull/69) Bump tox version from 3.27.0 to 3.27.1
 * [PR-72](https://github.com/pablobascunana/youml-api/pull/72) Rename repository from youml-manager to youml-api
 * [PR-68](https://github.com/pablobascunana/youml-api/pull/68) Refactor migrations