+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes will be documented in this file
 ## [0.0.12](https://github.com/pablobascunana/youml-api/compare/af06caf...develop) (WIP)
 
 ### Added
+* [PR-86](https://github.com/pablobascunana/youml-api/pull/86) Add train endpoint
 * [PR-85](https://github.com/pablobascunana/youml-api/pull/85) Encode and decode JWT
 * [PR-71](https://github.com/pablobascunana/youml-api/pull/71) Retrieve training endpoint
 * [PR-70](https://github.com/pablobascunana/youml-api/pull/70) Create training endpoint
@@ -13,6 +14,7 @@ All notable changes will be documented in this file
 * [PR-66](https://github.com/pablobascunana/youml-api/pull/66) Training model
 
 ### Changed
+* [PR-86](https://github.com/pablobascunana/youml-api/pull/86) Refactor mark-to-train endpoint
 * [PR-85](https://github.com/pablobascunana/youml-api/pull/85) Bump tox version from 4.0.11 to 4.0.18
 * [PR-84](https://github.com/pablobascunana/youml-api/pull/84) Rename some files references from youml-manager to youml-api
 * [PR-69](https://github.com/pablobascunana/youml-api/pull/69) Bump tox version from 3.27.0 to 3.27.1